+++ .github/CHANGELOG.md
@@ -26,6 +26,7 @@ This is the changelog for the version V2.6.0. Don't forget to [join us on Discor
 * #80 Engineers use missiles
 * massive rework on NodeEditor: Now completely on Server
 * count paths correctly
+* delte existing path now possible
 
 ### Bug fixes
 * #218 increased auto-sniper-time a little